fix: break all world.ts import chains from step-side modules
Introduce get-world-lazy.ts: a shared globalThis Symbol-based accessor
that replaces the static getWorld import in serialization.ts, run.ts,
helpers.ts, start.ts, and resume-hook.ts. This eliminates ALL static
import paths from step code to world.ts, keeping the step registrations
bundle small and free of process.cwd()/world-local/world-vercel code.

Also fix shouldPreferSourceBackedPackagePath in the deferred builder to
not prefer source-backed paths for tarball-installed packages in
node_modules, preventing the full source tree from being bundled.

Updated start.test.ts and step/writable-stream.test.ts to use
setWorld() instead of vi.mock for getWorld.
